+++ src/tabular/german_credit.jl
@@ -1,6 +1,12 @@
 struct GermanCredit <: TabularData end
 
-get_feature_names(data::GermanCredit) = get_feature_names("german_credit.csv")
+function get_original_feature_names(data::GermanCredit)
+    get_original_feature_names("german_credit.csv")
+end
+
+load_data(data::GermanCredit; kwrgs...) = load_german_credit(; kwrgs...)
+
+get_feature_names(data::GermanCredit) = load_data(data; feature_names=true)
 
 """
     load_german_credit(
@@ -17,6 +23,7 @@ function load_german_credit(
     seed=data_seed,
     train_test_split::Union{Nothing,Real}=nothing,
     shuffle::Bool=false,
+    kwrgs...,
 )
 
     # Setup:
@@ -34,7 +41,15 @@ function load_german_credit(
 
     # Pre-process:
     output = pre_process(
-        transformer, df_train, df_test; rng, nfinal_train, nfinal_test, ntotal, nreq
+        transformer,
+        df_train,
+        df_test;
+        rng,
+        nfinal_train,
+        nfinal_test,
+        ntotal,
+        nreq,
+        kwrgs...,
     )
 
     return output